Combine multiple interest functions into one common file and update references
