+++ src/main/common/interestUtils.ts
@@ -5,6 +5,8 @@ import { InterestType } from 'claim/form/models/interest'
 import { calculateInterest } from 'app/common/calculateInterest'
 import { Claim } from 'claims/models/claim'
 import { InterestData } from 'app/common/InterestData'
+import { ClaimAmountBreakdown } from 'claim/form/models/claimAmountBreakdown'
+import { DraftClaim } from 'drafts/models/draftClaim'
 
 export async function getInterestDetails (claim: Claim): Promise<InterestData> {
   if (claim.claimData.interest.type === InterestType.NO_INTEREST) {
@@ -26,4 +28,24 @@ export async function getInterestDetails (claim: Claim): Promise<InterestData> {
   const interestToDate: Moment = MomentFactory.currentDate().diff(interestFromDate, 'days') < 0 ? interestFromDate : MomentFactory.currentDate()
 
   return { numberOfDays, interest, rate, interestFromDate, interestToDate }
+}
+
+export async function draftInterestAmount (claimDraft: DraftClaim): Promise<number> {
+  const interest = claimDraft.interest
+  const breakdown: ClaimAmountBreakdown = claimDraft.amount
+  const interestDate = claimDraft.interestDate
+  const claimAmount: number = breakdown.totalAmount()
+
+  return calculateInterest(
+    claimAmount,
+    interest,
+    interestDate.type === InterestDateType.CUSTOM ? interestDate.date.toMoment() : MomentFactory.currentDate()
+  )
+}
+
+export async function draftClaimAmountWithInterest (claimDraft: DraftClaim): Promise<number> {
+  const interest: number = await draftInterestAmount(claimDraft)
+  const claimAmount: number = claimDraft.amount.totalAmount()
+
+  return claimAmount + interest
 }
