+++ src/main/features/claim/routes/pay.ts
@@ -12,7 +12,7 @@ import { CalculationOutcome } from 'app/fees/models/calculationOutcome'
 
 import { ClaimStoreClient } from 'app/claims/claimStoreClient'
 import { buildURL } from 'app/utils/callbackBuilder'
-import { claimAmountWithInterest } from 'app/utils/interestUtils'
+import { draftClaimAmountWithInterest } from 'common/interestUtils'
 import { User } from 'app/idam/user'
 import { DraftService } from 'services/draftService'
 import { ServiceAuthTokenFactoryImpl } from 'common/security/serviceTokenFactoryImpl'
@@ -78,7 +78,7 @@ export default express.Router()
       if (!draft.document.externalId) {
         throw new Error(`externalId is missing from the draft claim. User Id : ${user.id}`)
       }
-      const amount: number = await claimAmountWithInterest(draft.document)
+      const amount: number = await draftClaimAmountWithInterest(draft.document)
       if (!amount) {
         throw new Error('No amount entered, you cannot pay yet')
       }
